betty

+++ 0x1E-search_algorithms/0-linear.c
@@ -1,4 +1,5 @@
 #include "search_algos.h"
+
 /**
  * linear_search - searches for a value in an array of integers using the
  * Linear search algorithm
@@ -10,19 +11,19 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+    int i;
 
-	if (!array)
-		return (-1);
+    if (!array)
+        return (-1);
 
-	for (i = 0; i < size; i++)
-	{
-		printf("Value checked array[%lu] = [%d]\n", i, array[i]);
-		if (array[i] == value)
-		{
-			printf("Found %d at index: %lu\n", value, i);
-			return (i);
-		}
-	}
-	return (-1);
+    for (i = 0; i < (int)size; i++)
+    {
+        printf("Value checked array[%d] = [%d]\n", i, array[i]);
+        if (array[i] == value)
+        {
+            printf("Found %d at index: %d\n", value, i);
+            return (i);
+        }
+    }
+    return (-1);
 }
